Added a new function cleanData()
